NFDIV-386: refactored to a new controller/service and pointed to the new endpoint in CMS
