+++ build.gradle
@@ -37,7 +37,7 @@ def versions = [
         commonsIo                    : '2.6',
         commonsLang3                 : '3.9',
         commonsMath3                 : '3.6.1',
-        divCommonLib                 : '1.4.3',
+        divCommonLib                 : '1.4.4',
         dumbster                     : '1.7.1',
         elasticsearch                : '7.10.2',
         feignHttpClient              : '10.3.0',
@@ -110,6 +110,7 @@ allprojects {
     }
 
     repositories {
+        mavenLocal()
         mavenCentral()
         jcenter()
         maven { url "https://dl.bintray.com/hmcts/hmcts-maven" }
