+++ src/functionalTest/resources/fixtures/maintenance/submit/draft-divorce-session.json
@@ -1,3 +1,3 @@
 {
-  "DivorceOrDissolution": "divorce"
+  "divorceOrDissolution": "divorce"
 }
